refactor: separate JSON scenario construction
Move JSON translation into ScenarioJsonFactory and keep the core
builder and scenario construction path independent of JSON-specific
types. Preserve the traditional JSON workflow through the factory.
